requirements.txt
django-twitter-bootstrap
django-bootstrap3==4.7.0
requests
